Changed name of tapertc to cp-clockf83 and fixed various project files.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@30728 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/tapeport/tapelog.c
@@ -347,7 +347,7 @@ void tapelog_trigger_flux_change_passthrough(unsigned int on)
 {
     tapeport_trigger_flux_change(on, tapelog_device.id);
 
-    tapelog_transition("read", on);
+    tapelog_transition("read", (BYTE)on);
 
     tapelog_read = on;
 }
